requirements.txt
requests==2.27.1
python-dotenv==0.20.0
terminaltables==3.1.10
